Add new files manual/changes of types/fix issues -> (all examples fil…

+++ src/moduloD/d_array.h
@@ -7,8 +7,8 @@
 
 typedef struct {
     unsigned char *array;
-    int used;
-    int size;
+    size_t used;
+    size_t size;
 } D_Array;
 
 /**
